requirements_gpt.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.2
cycler==0.12.1
datasets==2.19.0
dill==0.3.8
distro==1.9.0
et-xmlfile==1.1.0
evaluate==0.4.2
exceptiongroup==1.2.0
filelock==3.14.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.6
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.3.2
jsonlines==4.0.0
kiwisolver==1.4.5
langid==1.1.6
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.16.2
openpyxl==3.1.3
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pdfminer.six==20221105
pdfplumber==0.10.3
pdfx==1.4.1
pillow==10.2.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.26.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.3
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sympy==1.12
tbb==2021.11.0
tokenizers==0.19.1
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.2
transformers==4.40.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
wsproto==1.2.0
xlwt==1.3.0
xxhash==3.4.1
yarl==1.9.4
